suit: Workaround for IPUC caches search before initialization
The IPUC cache was being searched for URI-s before it has been
initialized (before the first pull into the IPUC).

This led to issues - if only some images were already fetched and
the update was interrupted, the suit-processor would detect they
are already present in the cache. However, when pulling the
first missing image, the IPUC cache with the detected images
was erased leading to update failure.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>